requirements.txt
amazon-braket-algorithm-library==1.3.2
amazon-braket-default-simulator==1.12.1
amazon-braket-pennylane-plugin==1.15.1
amazon-braket-schemas==1.17.0
amazon-braket-sdk==1.39.0
antlr4-python3-runtime==4.9.2
anyio==3.7.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==23.1.0
autograd==1.6.2
autoray==0.6.3
awscli==1.27.65
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
boltons==23.0.0
boto3==1.26.64
botocore==1.29.65
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cirq-core==1.1.0
cirq-google==1.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.4
comm==0.1.3
contourpy==1.1.0
cvxpy==1.3.1
cycler==0.11.0
dask==2023.1.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.6
docutils==0.16
duet==0.2.8
ecos==2.0.12
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.17.1
fonttools==4.40.0
fqdn==1.5.1
fsspec==2023.6.0
future==0.18.3
google-api-core==1.34.0
google-auth==2.21.0
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.48.2
h5py==3.9.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.21.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jax==0.4.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.12.0
kiwisolver==1.4.4
locket==1.0.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mistune==3.0.1
mpmath==1.3.0
mypy-extensions==0.4.3
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==2.8.8
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.23.5
openfermion==1.5.1
openfermionpyscf==0.5
openpulse==0.4.2
openqasm3==0.4.0
opt-einsum==3.3.0
oqpy==0.1.1
osqp==0.6.3
overrides==7.3.1
packaging==23.1
pandas==2.0.0
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.0
pbr==5.11.1
PennyLane==0.30.0
PennyLane-Lightning==0.30.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
pluggy==1.2.0
ply==3.11
prometheus-client==0.17.0
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
PubChemPy==1.0.4
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
Pygments==2.15.1
pyparsing==3.1.0
pyrsistent==0.19.3
pyscf==2.2.1
pytest==7.4.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyYAML==5.4.1
pyzmq==25.1.0
qdldl==0.1.7
qiskit-aer==0.12.0
qiskit-braket-provider==0.0.4
qiskit-terra==0.23.3
qtconsole==5.4.3
QtPy==2.3.1
requests==2.31.0
retrying==1.3.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.7.2
rustworkx==0.13.0
s3transfer==0.6.1
scipy==1.9.3
scs==3.2.3
semantic-version==2.10.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
stack-data==0.6.2
stevedore==5.1.0
symengine==0.10.0
sympy==1.12
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
widgetsnbextension==4.0.7
